+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Privatize `Voice`, `VoiceLanguageInfo`, and `VoicesResponse` unnamed constructors.
 - Correct value type mapping of `VoicesResponse.toMap()`'s `voices` key.
 - Deprecate `VoicesResponse.getVoice()` method in favor of `VoicesResponse[]`.
+- Mark late properties of `Voice`, `VoiceLanguageInfo`, and `VoicesResponse` as getters.
 
 ## 1.0.0